% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airmass.R
\name{airmass}
\alias{airmass}
\title{Calculate air mass (AM) from zenith angle}
\usage{
airmass(z = 48.2)
}
\arguments{
\item{z}{zenith angle in degrees (default 48.2° corresponds to AM1.5)}
}
\value{
air mass value, numeric
}
\description{
Calculate the air mass value from the zenith angle
}
\details{
Calculate the air mass (AM) value using the secant approximation.
    Note that this function only returns physically meaningful values
    for z <= 70°,
}
\examples{
\dontrun{
    airmass()
    airmass(z = 59.83919) # latitude of chemistry dept at Ångström laboratory
}
}
\seealso{
https://en.wikipedia.org/wiki/Air_mass_(astronomy)
}
