% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airmass.R
\name{angzenith}
\alias{angzenith}
\title{Calculate zenith angle from air mass (AM)}
\usage{
angzenith(AM = 1.5)
}
\arguments{
\item{AM}{air mass, numeric (default 1.5)}
}
\value{
zenith angle, numeric
}
\description{
Calculate zenith angle corresponding to an air mass value
}
\details{
Calculate angle from the zenith line corresponding to an air mass value.
    This is the inverted function of \code{\link{airmass}}.
    Except I can't seem to figure out how to express it...
}
