% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airmass.R
\name{astrorefraction}
\alias{astrorefraction}
\title{Astronomical refraction}
\usage{
astrorefraction(z = 48.2)
}
\arguments{
\item{z}{zenith angle in degrees}
}
\value{
refraction, in degrees
    You might want to convert this value back to arcseconds using
    \code{\link[common]{degrees2asec}}
}
\description{
Calculate the astronomical refraction from zenith angle
}
\details{
Calculate the astronomical refraction, R,
    based on Oriani's power series in terms of tan Z (where Z is
    the zenith angle) neglecting terms beyond third order and under
    the assumption that the height of the atmosphere is small compared
    to the radius of the Earth.
    This function works well up to around 70° of zenith angle, but will
    return non-physical values at higher angles.
    Astronomical refraction becomes complicated close to the horizon (for more
    on that, see quoted sources).
}
\seealso{
Mahan, A.I., 1962. Astronomical refraction–some history and theories. Appl. Opt., AO 1, 497–511. https://doi.org/10.1364/AO.1.000497

Woolard, E.W., Clemence, G.M., 1966. Spherical astronomy. Elsevier.

Auer, L.H., Standish, E.M., 2000. Astronomical refraction: computational method for all zenith angles. AJ 119, 2472. https://doi.org/10.1086/301325
}
