% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux.R
\name{cumflux}
\alias{cumflux}
\title{Total photon flux across one ore more wavelength ranges}
\usage{
cumflux(wl.start = 280, wl.stop = 4000, model = "AM1.5G")
}
\arguments{
\item{wl.start}{numeric vector, in nanometres}

\item{wl.stop}{numeric vector, in nanometres}

\item{model}{string, defaults to AM1.5G. Options are: AM0, AM1.5G, or DNCS}
}
\value{
the total flux across the wl range, number (vector if more than one range)
}
\description{
This function returns the total (cumulative) flux across specified
wavelength range(s), i.e., photons per second per square metre.
NOTE, this function supports a shorthand argument style:
if only *one* argument is supplied, that argument will be treated
as the range maxima (and the function will set all range minima
to 280 nm for you).
}
\examples{
\dontrun{
cumflux()
cumflux(wl.start = c(280, 400), c(400, 700))
cumflux(400, 700, model = "DNCS")
cumflux(wl.stop = c(1000, 2000, 3000, 4000))
# the following short-hand notation also works,
# (wl.start is assumed to be 280 nm):
cumflux(c(1000, 2000, 3000, 4000))
}
}
