% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentdensity.R
\name{currentdensity}
\alias{currentdensity}
\title{Calculate short-circuit current density}
\usage{
currentdensity(energy, photonflux, EQE = 1, bandgap = 1.23)
}
\arguments{
\item{energy, }{energy/eV, numeric vector}

\item{photonflux, }{cumulative photon flux/s⁻¹ m⁻², vector}

\item{EQE, }{external quantum efficiency, numeric (0 - 1)}

\item{bandgap, }{band gap of photoabsorber, numeric/eV (equivalent to electrode potential)}
}
\value{
dataframe with 3 columns:
   \describe{
      \item{energy}{energy/eV, same as input vector}
      \item{currentdensity}{current density/mA cm⁻² (or equivalently, A m⁻²)}
      \item{currentdensity.fraction}{relative current density}
   }
}
\description{
Note that for the sake of following convention this function
   uses energy (in eV) rather than wavelength as its x-axis.
}
\details{
Calculate short-circuit (maximum) current density (Jsc) from the photon flux.
}
\examples{
\dontrun{
   currentdensity(
      energy = photoec::sunlight.ASTM(model="AM1.5G")$energy,
      photonflux = photoec::sunlight.ASTM(model="AM1.5G")[["AM1.5G.photonflux"]]
   )
}
}
