% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit-converters.R
\name{energy2wavelength}
\alias{energy2wavelength}
\title{energy2wavelength}
\usage{
energy2wavelength(eV)
}
\arguments{
\item{eV}{vector or scalar, in eV}
}
\value{
nanometers
}
\description{
Convert from electron volt to nanometer
}
\examples{
\dontrun{
energy2wavelength(3.25)
energy2wavelength(1.92)
}
}
