% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit-converters.R
\name{energy2wavenum}
\alias{energy2wavenum}
\title{Convert energy to wavenumber}
\usage{
energy2wavenum(energy)
}
\arguments{
\item{energy}{numeric, scalar or vector}
}
\value{
numeric, scalar or vector
}
\description{
Converts energy (eV) to wavenumber (cm-1)
}
