% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{eV2nm}
\alias{eV2nm}
\title{eV2nm}
\usage{
eV2nm(eV)
}
\arguments{
\item{eV}{vector or scalar, in eV}
}
\value{
nanometers
}
\description{
Convert from electron volt to nanometer
DEPRECATED. Please use energy2wavelength() instead.
}
