% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux.R
\name{flux}
\alias{flux}
\title{Photon flux at a particular wavelength}
\usage{
flux(wavelength, model = "AM1.5G")
}
\arguments{
\item{wavelength}{One or more wavelength values, in nanometres}

\item{model}{defaults to AM1.5G. Options are: AM0, AM1.5G, or DNCS}
}
\value{
Vector with flux values at the given wavelength values.
}
\description{
Note that this function returns the flux at the particular wavelength,
i.e., photons per second per square metre per nanometre.
It happily accepts a vector of wavelength, returning a vector of fluxes
at each particular wavelength (same as before).
To get the total photon flux across a wavelength range use the
function \code{flux.total}
}
