% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photonflux.R
\name{photonflux}
\alias{photonflux}
\title{Calculate the photon flux}
\usage{
photonflux(wavelength, spectralirradiance)
}
\arguments{
\item{wavelength, }{usually in nm}

\item{spectralirradiance, }{usually in W m-2 nm-1}
}
\value{
Dataframe with 4 columns:
    \describe{
       \item{photonflux}{Photons per unit area, photon flux / s-1 m-2 nm-1}
       \item{photonflux.trapz}{numerically integrated}
       \item{photonflux.csum}{cumulative photon flux, s-1 m-2}
       \item{photonflux.fraction}{photon flux fraction (from 0 - 1)}
}
}
\description{
Calculate photon flux from spectral irradiance
}
\details{
Provide wavelength and spectral irradiance.
    It is important that provided wavelength and spectral irradiance
    use the same length unit (for example, nm and W m-2 nm-1, or m and W m-2 m-1).
}
