% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STH.R
\name{STH}
\alias{STH}
\title{Calculate maximum solar-to-hydrogen efficiency}
\usage{
STH(
  energy,
  currentdensity,
  max.irradiance = 1000.371,
  model = "AM1.5G",
  electrode.potential = 1.23,
  faradaic.efficiency = 1
)
}
\arguments{
\item{energy, }{energy/eV, numeric vector}

\item{currentdensity, }{current density/mA cm⁻²}

\item{max.irradiance, }{total irradiance (for specified model)/W m⁻², numeric scalar.
Note that `max.irradiance` only takes precedence if model = ""}

\item{model, }{ASTM model: "AM1.5G" (default), "AM0", or "DNCS"}

\item{electrode.potential, }{electrode potential/eV, numeric scalar}

\item{faradaic.efficiency}{Faradaic efficiency, numeric (0 - 1)}
}
\value{
dataframe with 2 columns:
    \describe{
       \item{energy}{energy/eV, same as input vector}
       \item{STH}{STH/\%}
    }
}
\description{
Calculate the maximum STH efficiency
}
\examples{
\dontrun{
   STH(
      energy = photoec::sunlight.ASTM(model = model)$energy,
      currentdensity = photoec::sunlight.ASTM(model = model)[[paste0(model, ".currentdensity")]]
   )
}
}
