% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunlightASTM.R
\name{sunlight.ASTM}
\alias{sunlight.ASTM}
\title{Calculate irradiances based on ASTM G173-03 reference}
\usage{
sunlight.ASTM(
  wavelength = c(seq(280, 400, 0.5), seq(401, 1700, 1), 1702, seq(1705, 4000, 5)),
  model = ""
)
}
\arguments{
\item{wavelength}{defaults to the wavelength values used in the G173-03 model}

\item{model}{accepts empty string (default), or "AM1.5G", "AM0", or "DNCS"}
}
\value{
NOTE: which model is returned depends on the "model" argument
    \describe{
       \item{wavelength}{Wavelength/nm}
       \item{<model>.spectralirradiance}{Spectral irradiance/W m⁻² nm⁻¹}
       \item{<model>.irradiance}{Cumulative irradiance/W m⁻²}
       \item{<model>.irradiance.fraction}{Relative irradiance (0-1)}
       \item{<model>.spectralphotonflux}{Spectral photon flux/s⁻¹ m⁻² nm⁻¹}
       \item{<model>.photonflux}{Cumulative photon flux/s⁻¹ m⁻²}
       \item{<model>.photonflux.fraction}{Relative photon flux (0-1)}
    }
}
\description{
Calculate the spectral irradiance for sunlight at or below
    Earth's atmosphere based on ASTM G173-03 reference spectra
}
\details{
Calculate the spectral irradiance of sunlight on Earth
    based on ASTM G173-03 model for any wavelength between
    280 nm and 4000 nm with at best 0.01 nm resolution (interpolated).
}
\examples{
\dontrun{
    sunlight.ASTM()
    sunlight.ASTM(model = "AM0")
    sunlight.ASTM(wavelength = seq(400, 450, by = 0.1), model = "AM1.5G")
    sunlight.ASTM(wavelength = seq(500, 600, by = 1))
}
}
