% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit-converters.R
\name{wavelength2energy}
\alias{wavelength2energy}
\title{wavelength2energy}
\usage{
wavelength2energy(nm)
}
\arguments{
\item{nm}{vector or scalar, in nm}
}
\value{
electron volts
}
\description{
Convert from nanometer to electron volt
}
