% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit-converters.R
\name{wavenum2energy}
\alias{wavenum2energy}
\title{Convert wavenumber to energy}
\usage{
wavenum2energy(wavenumber)
}
\arguments{
\item{wavenumber}{numeric, scalar or vector}
}
\value{
numeric, scalar or vector
}
\description{
Converts wavenumber (cm-1) to energy (eV)
}
