% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{cumflux}
\alias{cumflux}
\title{Total photon flux across one ore more wavelength ranges}
\usage{
cumflux(wl.start = 280, wl.stop = 4000, model = "AM1.5G")
}
\arguments{
\item{wl.start}{numeric vector, in nanometres}

\item{wl.stop}{numeric vector, in nanometres}

\item{model}{string, defaults to AM1.5G. Options are: AM0, AM1.5G, or DNCS}
}
\value{
the total flux across the wl range, number (vector if more than one range)
}
\description{
DEPRECATED. please use \link{sunlight.ASTM()} instead.
}
