% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentdensity.R
\name{currentdensity}
\alias{currentdensity}
\title{Calculate short-circuit current density}
\usage{
currentdensity(
  wavelength,
  photonflux.csum,
  quantum.efficiency = 1,
  bandgap = 1.23
)
}
\arguments{
\item{wavelength, }{in nanometer}

\item{photonflux.csum, }{s-1 m-2}

\item{quantum.efficiency, }{between 0 and 1}

\item{bandgap, }{in eV (equivalent to electrode potential)}
}
\value{
dataframe: energy (eV), wavelength (nm), and current density (A m-2)
}
\description{
Calculate the short-circuit current density (Jsc) from wavelength and photon flux.
}
\details{
If you use this function together with STH(), set the 
  quantum.efficiency only once, either here or in \code{\link{STH}}.
}
