% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{flux}
\alias{flux}
\title{Photon flux at a particular wavelength}
\usage{
flux(wavelength, model = "AM1.5G")
}
\arguments{
\item{wavelength}{One or more wavelength values, in nanometres}

\item{model}{defaults to AM1.5G. Options are: AM0, AM1.5G, or DNCS}
}
\value{
Vector with flux values at the given wavelength values.
}
\description{
DEPRECATED. Please use \link{sunlight.ASTM()} instead.
}
