% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STH.R
\name{STH}
\alias{STH}
\title{Calculate maximium solar-to-hydrogen efficiency}
\usage{
STH(wavelength, currentdensity, total.irradiance = max(sunlight.ASTM(model =
  "AM1.5G")$irradiance), electrode.potential = 1.23,
  faradaic.efficiency = 1)
}
\arguments{
\item{wavelength, }{in nanometer}

\item{currentdensity, }{in A m-2}

\item{total.irradiance, }{in W m-2 (note: scalar, not vector)}

\item{electrode.potential, }{in eV (note: scalar, not vector)}

\item{faradaic.efficiency}{scalar, not vector}
}
\value{
dataframe with 3 columns: : energy (eV), wavelength (nm), and current density (A m-2)
}
\description{
Calculate the maximum STH efficiency
}
