% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunlightASTM.R
\name{sunlight.ASTM}
\alias{sunlight.ASTM}
\title{Calculate irradiances based on ASTM G173-03 reference}
\usage{
sunlight.ASTM(wavelength = c(seq(280, 400, 0.5), seq(401, 1700, 1), 1702,
  seq(1705, 4000, 5)), model = "")
}
\arguments{
\item{wavelength}{defaults to the wavelength values used in the G173-03 model}

\item{model}{defaults to empty string, accepts "AM1.5G", "AM0", or "DNCS"}
}
\value{
NOTE: which columns are returned depends on the argument model = ...
    \describe{
       \item{wavelength}{Wavelength, in nm}
       \item{AM0.spectralirradiance}{W m-2 nm-1}
       \item{AM1.5G.spectralirradiance}{W m-2 nm-1}
       \item{DNCS.spectralirradiance}{W m-2 nm-1}
       \item{AM0.spectralirradiance.trapz}{W m-2}
       \item{AM1.5G.spectralirradiance.trapz}{W m-2}
       \item{DNCS.spectralirradiance.trapz}{W m-2}
       \item{AM0.irradiance}{W m-2}
       \item{AM1.5G.irradiance}{W m-2}
       \item{DNCS.irradiance}{W m-2}
       \item{AM0.irradiance.fraction}{fraction}
       \item{AM1.5G.irradiance.fraction}{fraction}
       \item{DNCS.irradiance.fraction}{fraction}
    }
}
\description{
Calculate the spectral irradiance for sunlight at or below
    Earth's atmosphere based on ASTM G173-03 reference spectra
}
\details{
Calculate the spectral irradiance of sunlight on Earth
    based on ASTM G173-03 model for any wavelength between
    280 nm and 4000 nm with at best 0.01 nm resolution (interpolated).
}
\examples{
\dontrun{
sunlight.ASTM(wavelength = seq(400, 450, by = 0.1), model = "AM1.5G")
}
}
