% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunlightPlanck.R
\name{sunlight.Planck}
\alias{sunlight.Planck}
\title{Calculate spectral radiance and other characteristics of a black-body}
\usage{
sunlight.Planck(wavelength, temperature = 5778)
}
\arguments{
\item{wavelength}{vector, in meter}

\item{temperature}{vector, blackbody surface temperature, in Kelvin}
}
\value{
dataframe
}
\description{
Calculate the spectral radiance for an ideal black-body
    using Planck's law, for a specified wavelength range and step-size
}

